CircleCI and User Profiles (#218)
* Authentication setup

* CI/CD setup with CircleCI

+++ install_gitcrypt.sh
@@ -1,7 +1,9 @@
-#!/bin/bash
+#!/bin/sh
 
 cd /tmp/
-sudo rm -r *
+if [ "$(ls -A /tmp)" ]; then
+    sudo rm -r *
+fi
 git clone https://github.com/AGWA/git-crypt.git /tmp/git-crypt
 cd /tmp/git-crypt
 make
